PRSD-NONE: Moves JourneyDataExtensions into journeyDataExtensions sub-folder
